+++ examples/typescript/http-server-pool/fastify-cluster/src/main.ts
@@ -19,13 +19,13 @@ const pool = new FixedClusterPool<WorkerData, WorkerResponse>(
           if (response.status) {
             console.info(
               // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-              `Fastify is listening on worker on port ${response.port}`
+              `Fastify is listening on cluster worker on port ${response.port}`
             )
           }
           return null
         })
         .catch(error => {
-          console.error('Fastify failed to start on worker:', error)
+          console.error('Fastify failed to start on cluster worker:', error)
         })
     },
     errorHandler: (e: Error) => {