+++ examples/typescript/websocket-server-pool/ws-hybrid/src/main.ts
@@ -1,16 +1,16 @@
 import { dirname, extname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { FixedClusterPool, availableParallelism } from 'poolifier'
-import type { ClusterWorkerData, ClusterWorkerResponse } from './types.js'
+import { type ClusterWorkerData, type ClusterWorkerResponse } from './types.js'
 
-const fastifyWorkerFile = join(
+const webSocketServerWorkerFile = join(
   dirname(fileURLToPath(import.meta.url)),
-  `fastify-worker${extname(fileURLToPath(import.meta.url))}`
+  `websocket-server-worker${extname(fileURLToPath(import.meta.url))}`
 )
 
 const pool = new FixedClusterPool<ClusterWorkerData, ClusterWorkerResponse>(
   Math.round(availableParallelism() / 2),
-  fastifyWorkerFile,
+  webSocketServerWorkerFile,
   {
     onlineHandler: () => {
       pool
@@ -19,17 +19,20 @@ const pool = new FixedClusterPool<ClusterWorkerData, ClusterWorkerResponse>(
           if (response.status) {
             console.info(
               // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-              `Fastify is listening on cluster worker on port ${response.port}`
+              `WebSocket server is listening on cluster worker on port ${response.port}`
             )
           }
           return null
         })
         .catch(error => {
-          console.error('Fastify failed to start on cluster worker:', error)
+          console.error(
+            'WebSocket server failed to start on cluster worker:',
+            error
+          )
         })
     },
     errorHandler: (e: Error) => {
-      console.error('Cluster worker error:', e)
+      console.error('Cluster worker error', e)
     }
   }
 )